+++ Assets/Mirage/Runtime/Events.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 267a30192d7c133408ba0478f68e0d11
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
