+++ Assets/Mirage/Runtime/Events/IRunOnceEvent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 483cef5299abdc5409cfe4d951692a06
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
