+++ Assets/Mirage/Runtime/Events/RunOnceEvent.cs
@@ -2,41 +2,8 @@ using System;
 using UnityEngine;
 using UnityEngine.Events;
 
-namespace Mirage
+namespace Mirage.Events
 {
-    public abstract class RunOnceEventBase
-    {
-        protected abstract UnityEventBase baseEvent { get; }
-        protected bool hasInvoked { get; private set; }
-
-        protected void MarkInvoked()
-        {
-            if (hasInvoked) throw new InvalidOperationException("Event can only be invoked once Invoke");
-
-            hasInvoked = true;
-        }
-
-        /// <summary>
-        /// Resets event, removing all listens and allowing it to be invoked again
-        /// </summary>
-        public void Reset()
-        {
-            hasInvoked = false;
-            baseEvent.RemoveAllListeners();
-        }
-    }
-
-    /// <summary>
-    /// Event that can only run once, adding handler late will it invoke right away
-    /// </summary>
-    /// <remarks>
-    /// Interface only contains AddHandler method because Invoke should only be called from the owner of the event
-    /// </remarks>
-    public interface IRunOnceEvent
-    {
-        void AddListener(UnityAction handler);
-    }
-
     /// <summary>
     /// Event that can only run once, adding handler late will it invoke right away
     /// </summary>
@@ -65,14 +32,6 @@ namespace Mirage
 
             _event.Invoke();
         }
-    }
-
-    /// <summary>
-    /// Version of <see cref="IRunOnceEvent"/> with 1 argument
-    /// </summary>
-    public interface IRunOnceEvent<T0>
-    {
-        void AddListener(UnityAction<T0> handler);
     }
 
     /// <summary>
@@ -109,14 +68,6 @@ namespace Mirage
             this.arg0 = arg0;
             _event.Invoke(arg0);
         }
-    }
-
-    /// <summary>
-    /// Version of <see cref="IRunOnceEvent"/> with 2 arguments
-    /// </summary>
-    public interface IRunOnceEvent<T0, T1>
-    {
-        void AddListener(UnityAction<T0, T1> handler);
     }
 
     /// <summary>
