+++ Assets/Mirage/Runtime/Events/RunOnceEventBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e5e161bb849a4ff4ab541c3099513315
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
